Добавлены проверки bouncer

+++ app/Controllers/Http/StaffController.ts
@@ -3,7 +3,13 @@ import Staff from 'App/Models/Staff'
 import StaffValidator from 'App/Validators/StaffValidator'
 
 export default class StaffController {
-  public async index({ request, view }: HttpContextContract) {
+  public async index({ request, response, view, session, bouncer }: HttpContextContract) {
+    if (await bouncer.denies('noAccess')) {
+      session.flash('dangerMessage', 'У вас нет доступа к данному разделу!')
+
+      return response.redirect().toPath('/')
+    }
+
     const page = request.input('page', 1)
     const limit = 10
     const staff = await Staff.query().orderBy('created_at', 'asc').paginate(page, limit)
@@ -17,7 +23,13 @@ export default class StaffController {
     })
   }
 
-  public async create({ view }: HttpContextContract) {
+  public async create({ response, view, session, bouncer }: HttpContextContract) {
+    if (await bouncer.denies('noAccess')) {
+      session.flash('dangerMessage', 'У вас нет прав на добавление новой записи!')
+
+      return response.redirect().toPath('/')
+    }
+
     return view.render('pages/staff/form', {
       title: 'Добавление нового сотрудника',
       options: {
@@ -28,7 +40,13 @@ export default class StaffController {
     })
   }
 
-  public async store({ request, response, session }: HttpContextContract) {
+  public async store({ request, response, session, bouncer }: HttpContextContract) {
+    if (await bouncer.denies('noAccess')) {
+      session.flash('dangerMessage', 'У вас нет прав на добавление новой записи!')
+
+      return response.redirect().toPath('/')
+    }
+
     const validateStaffData = await request.validate(StaffValidator)
 
     if (validateStaffData) {
@@ -47,8 +65,14 @@ export default class StaffController {
 
   public async show({}: HttpContextContract) {}
 
-  public async edit({ params, response, view, session }: HttpContextContract) {
+  public async edit({ params, response, view, session, bouncer }: HttpContextContract) {
     const employee = await Staff.find(params.id)
+
+    if (await bouncer.denies('noAccess')) {
+      session.flash('dangerMessage', 'У вас нет прав на редактирование записи!')
+
+      return response.redirect().toPath('/')
+    }
 
     if (employee) {
       return view.render('pages/staff/form', {
@@ -67,8 +91,14 @@ export default class StaffController {
     }
   }
 
-  public async update({ params, request, response, session }: HttpContextContract) {
+  public async update({ params, request, response, session, bouncer }: HttpContextContract) {
     let employee = await Staff.find(params.id)
+
+    if (await bouncer.denies('noAccess')) {
+      session.flash('dangerMessage', 'У вас нет прав на редактирование записи!')
+
+      return response.redirect().toPath('/')
+    }
 
     if (employee) {
       const validateStaffData = await request.validate(StaffValidator)
@@ -88,8 +118,14 @@ export default class StaffController {
     }
   }
 
-  public async destroy({ response, params, session }: HttpContextContract) {
+  public async destroy({ response, params, session, bouncer }: HttpContextContract) {
     const employee = await Staff.find(params.id)
+
+    if (await bouncer.denies('noAccess')) {
+      session.flash('dangerMessage', 'У вас нет прав на удаление записи!')
+
+      return response.redirect().toPath('/')
+    }
 
     if (employee) {
       await employee.delete()
